requirements.txt
fastapi==0.115.6
pydantic==2.10.6
PyJWT==2.10.1

